fix: Close the branch and archive the left over due to too many cases and hard to maintain
This commit closes the branch and archives the left over code due to the fact that the approach used in the rewrite version did not work well and had too many cases that needed to be handled individually, making it difficult to maintain in the future. Despite this, some techniques from this approach will still be used in the new approach.