Init: This an early look how v2 gonna be. I just finish design how the keyword would be store, how the tokenizer would be.
For the new version, we gonna use possibility based approach. So, first we tokenize them. Then, we assign what the possibility each token might be. The next part is we score each possibility based on neighbour token. Once everything is scored, we iterate each token and pick which possibility each token would be based on the score.

When choosing the token possibility, it might do a look up to ensure some category that not conflict, for example, we cant have 2 crc32 value in a file. After that, we iterate through all token that dont have any possibility, and we match using known pattern.