ci: Output standardisation
This project kind of dead end to me right now, While this have solid structure. It just too much to code and considered.

TODO:
A new tokenization approach to make use of keyword/db for the knowledge.
Re-evaluate the Token to ensure we can handle overlaped char to multiple token, `s` in `10bitsubs`.
Re-evaluate how to store token possibilites so we can easily use simple version for scoring, and detailed version for the output.
Tuning the scoring system.

Plan:
As today, I will try to look how to utilized local LLM instead.